# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/baskets/{paramBasketId}/products/{paramProductId}": {
            "post": {
                "tags": [
                    "Basket"
                ],
                "operationId": "addItem",
                "parameters": [
                    {
                        "name": "paramBasketId",
                        "in": "path",
                        "description": "The basket's id",
                        "required": true,
                        "example": "2"
                    },
                    {
                        "name": "paramProductId",
                        "in": "path",
                        "description": "The product's id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": ""
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item added to basket.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                {
                                                    "NO_BASKET": "The basket doesn't exist or doesn't belong to the authenticated user."
                                                },
                                                {
                                                    "CLOSED_BASKET": "The basket is checked out, it can no longer be modified."
                                                },
                                                {
                                                    "BAD_PRODUCT": "The product does not exist."
                                                },
                                                {
                                                    "PRODUCT_IS_ALREADY_ADDED": "The basket already contains that product."
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Access denied.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Basket"
                ],
                "operationId": "removeItem",
                "parameters": [
                    {
                        "name": "paramBasketId",
                        "in": "path",
                        "description": "The basket's id",
                        "required": true,
                        "example": "2"
                    },
                    {
                        "name": "paramProductId",
                        "in": "path",
                        "description": "The product's id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": ""
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item removed from basket.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                {
                                                    "NO_BASKET": "The basket doesn't exist or doesn't belong to the authenticated user."
                                                },
                                                {
                                                    "CLOSED_BASKET": "The basket is checked out, it can no longer be modified."
                                                },
                                                {
                                                    "BAD_PRODUCT": "The product does not exist."
                                                },
                                                {
                                                    "PRODUCT_IS_NOT_AVAILABLE": "The basket doesn't contain that product."
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Access denied.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/products/download-losses": {
            "get": {
                "tags": [
                    "Product"
                ],
                "operationId": "downloadLosses",
                "responses": {
                    "200": {
                        "description": "Item added to basket.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Access denied.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}